WAHElements

API: WAHElements

WAHVectors.isruns — Method.

isruns(x::WAHElement)

Return true if the element of the WAH compressed bit array x represents a compressed run of words. Otherwise return false.

source

WAHVectors.isliteral — Method.

isliteral(x::WAHElement)

Return true if the element of the WAH compressed bit array x represents a literal word rather than a compressed run of words, in which case this function will return false.

source

WAHVectors.is_zeros_runs — Method.

is_zeros_runs(x::WAHElement)

Return true if the WAH Element represents a compressed run of all zero words.

source

WAHVectors.is_ones_runs — Method.

is_ones_runs(x::WAHElement)

Return true if the WAH Element represents a compressed run of all one words.

source

WAHVectors.nwords — Method.

nwords(x::WAHElement)

Get the number of words represented by this single WAHElement.

source

WAHVectors.runval — Method.

runval(x::WAHElement)

Get the runvalue of a compressed WAHElement.

Warning

Output will not make sense if you use it on a WAH element that represents a literal word (i.e. isliteral(x) returns true) rather than a number of compressed words.

source

WAHVectors.nruns — Method.

nruns(x::WAHElement)

Get the number of words represented by this single WAHElement, assuming it is a compressed element.

Warning

Output will not make sense if you use it on a WAH element that represents a literal word (i.e. isliteral(x) returns true) rather than a number of compressed words.

source

WAHVectors.nfree — Method.

nfree(x::WAHElement)

Return how many more words this WAHElement can compress into it.

Warning

Output will not make sense if you use it on a WAH element that represents a literal word (i.e. isliteral(x) returns true) rather than a number of compressed words.

source

WAHVectors.isfull — Method.

isfull(x::WAHElement)

Check whether this WAHElement contains as many words as it can.

Warning

Output will not make sense if you use it on a WAH element that represents a literal word (i.e. isliteral(x) returns true) rather than a number of compressed words.

source